+++ packages/core/rust/types.d.ts
@@ -234,6 +234,7 @@ export class AtlaspackTracer {
   constructor();
   enter(label: string): SpanId;
   exit(id: SpanId): void;
+  record(id: SpanId, data?: any | undefined | null): void;
 }
 export class Resolver {
   constructor(projectRoot: string, options: FileSystem);
@@ -378,14 +379,10 @@ export class SourceMap {
   getProjectRoot(): string;
 }
 
-export declare function atlaspackNapiLoadBundleGraph(
-  atlaspackNapi: AtlaspackNapi,
-  nodes: Array<BundleGraphNode>,
-  edges: Array<[number, number]>,
-): object;
 export declare function atlaspackNapiPackage(
   atlaspackNapi: AtlaspackNapi,
-): object;
+  bundleId: string,
+): Promise<[RunPackagerRunnerResult, AtlaspackNapiError]>;
 export interface CompiledCssInJsConfigPlugin {
   configPath?: string;
   importReact?: boolean;
@@ -419,3 +416,10 @@ export interface CompiledCssInJsPluginResult {
   diagnostics: Array<JsDiagnostic>;
   bailOut: boolean;
 }
+export declare function atlaspackNapiLoadBundleGraph(
+  atlaspackNapi: AtlaspackNapi,
+  nodes: string,
+  edges: Array<[number, number, number]>,
+  publicIdByAssetId: Record<string, string>,
+  environments: string,
+): object;