+++ packages/validators/typescript/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaspack/validator-typescript
 
+## 2.14.50
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/utils@3.3.2
+  - @atlaspack/plugin@2.14.50
+  - @atlaspack/types@2.15.40
+  - @atlaspack/ts-utils@2.14.29
+
 ## 2.14.49
 
 ### Patch Changes